requirements.txt
qgiscommons
owslib

# test requirements